% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evalPsi.R
\name{evalPsi}
\alias{evalPsi}
\title{Evaluate covariance matrix for kriging}
\usage{
evalPsi(DTR, L, lamb.est, theta, PhiTime, PhiTimeTE = NULL, homogeneous,
  subsetStatic, kriging = FALSE)
}
\arguments{
\item{DTR}{Matrix of distances}

\item{L}{Number of functional principal components.}

\item{lamb.est}{Vector of eigenvalues.}

\item{theta}{Vector of spatial dependence parameters.}

\item{PhiTime}{Vector of times, in numeric mode.}

\item{PhiTimeTE}{Either NULL (repeats PhiTime) or vector of times,
in numeric mode (for prediction)}

\item{homogeneous}{Logical flag for homogeneous model.}

\item{subsetStatic}{Logical vector for subset of static sensors}

\item{kriging}{Logical flag, whether the function is evaluating a covariance
               matrix (FALSE) or a kriging cross-covariance matrix (TRUE).
               Defaults to FALSE.}
}
\value{
List of one items
  \item{psi.cov}{Desired covariance matrix}
}
\description{
This function is for internal use on the STDF algorithm. It evaluates the
covariance matrix for fitting, and the kriging matrix for prediction.
}
\examples{
\dontrun{
# Syntax example
stdf::evalPsi(DTR = matrix(c(1,0,0,1), 2),
        L = 1, lamb.est = 1,
        theta = 1:2, PhiTime = matrix(1:2, ncol=1), PhiTimeTE = NULL,
        homogeneous = TRUE, subsetStatic = rep(1, 2))
        }
}
\seealso{
\code{\link{stdf}}
}
\keyword{Spatial}
\keyword{Statistics}

