% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gfda.R
\name{gfda}
\alias{gfda}
\title{GFDA model}
\usage{
gfda(training.set, prediction.set, subtfpca = NULL, ssensors = 6, J = 2,
  spline.df = NULL, fpca.df = 20, homogeneous = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{training.set}{A numeric matrix with the response variable in the first column,
time values in the second column and X, Y coordinates in the third
column. Corresponds to the training data.}

\item{prediction.set}{Same as training.set, but corresponding to points to be predicted for
MSPE calculation.}

\item{ssensors}{Integer corresponding to how many static sensors in the dataset.}

\item{J}{Number of eigenfunctions in spatio-temporal covariance, defaults to 2.}

\item{spline.df}{Number of spline basis for the deterministic spline component (see \code{\link{bs}} function).}

\item{fpca.df}{Number of spline basis for the stochastic spline component (see \code{\link{tfpca}} function).}

\item{homogeneous}{Whether the variance of static and roving sensors is assumed to be the same or not. Defaults to not.}

\item{verbose}{Wheter \code{gfda} gives detailed information on optimization step, defaults to \code{TRUE}.}
}
\value{
List of four items
  \item{beta.est}{Coefficient estimates}
  \item{MSPEKrig}{Mean squared Kriging Prediction error}
  \item{spatCov}{Theta parameters for stochastic term}
  \item{sigma2s}{Static sensor variability}
  \item{sigma2r}{Roving sensor variability}
}
\description{
This function is an wrapper to fitting a Geostatistical Functional Data
Analysis (GFDA) model to spatio-temporal data.
}
\examples{
## require(fda)
## Y <- CanadianWeather$dailyAv
## XY <- CanadianWeather$coordinates
# static sensor simulation example:
set.seed(1)
X <- cbind(rnorm(200), rep(1:50,4), rep(c(2,2,4,4), each = 50), rep(c(2,4,2,4), each = 50))
Xpred <- cbind(rnorm(50), 1:50, 3, 3)
results <- gfda(X, Xpred, ssensors = 4)
resultsH <- gfda(X, Xpred, ssensors = 4, homogeneous = TRUE)
X2 <- rbind(X, cbind(rnorm(50), 1:50, seq(2,4, length=50), seq(2,4, length=50)))
resultsPlusRoving <- gfda(X2, Xpred, subtfpca = c(rep(TRUE,200), rep(FALSE,50)), ssensors = 4)
}
\references{
\url{http://www.google.com}
}
\seealso{
\code{\link{tfpca}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

