% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stdf}
\alias{predict.stdf}
\title{Predict method STDF Model Fits}
\usage{
\method{predict}{stdf}(object, newdata = NULL, what = c("fit", "predict",
  "loadings"), se = FALSE, loadings.range = list(x = c(0, 1), y = c(0, 1),
  mesh = 20), ...)
}
\description{
Predicted values for a \code{stdf} object.
}

