% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stdf.R
\name{stdf}
\alias{stdf}
\title{STDF model}
\usage{
stdf(training.set, prediction.set, subtfpca = NULL, ssensors = 6, L = 2,
  spline.df = NULL, fpca.lambda = 0, fpca.df = 20, homogeneous = FALSE,
  method = "L-BFGS-B", verbose = TRUE, ...)
}
\arguments{
\item{training.set}{A numeric matrix with the response variable in the
first column, time values in the second column and X, Y
spatial coordinates in the third column. Corresponds to
the training data.}

\item{prediction.set}{Same as training.set, but corresponding to points to
be predicted for MSPE calculation.}

\item{subtfpca}{Logical vector with the same length as the number of rows in
training.set; tells stdf which lines should be included in
the calculation of the temporal dependency component.
Defaults to NULL, which includes all lines.}

\item{ssensors}{Integer corresponding to how many static sensors in the
dataset.}

\item{L}{Number of eigenfunctions in spatio-temporal covariance, defaults
to 2.}

\item{spline.df}{Number of spline basis for the deterministic spline
component (see \code{\link{bs}} function). Defaults to
NULL, which uses 3 spline basis.}

\item{fpca.lambda}{Smoothness parameter for the fpca function. Defaults
to 0 (no smoothing). See \code{\link{tfpca}} function.}

\item{fpca.df}{Number of spline basis for the stochastic spline component
(see \code{\link{tfpca}} function). Defaults to 20.}

\item{homogeneous}{Whether the variance of static and roving sensors is
assumed to be the same or not. Defaults to FALSE.}

\item{method}{Either "L-BFGS-B" or "Nelder-Mead", which are passed to the
              optimization function \code{\link{constrOptim}}.}

\item{verbose}{Whether \code{stdf} prints a message when the optimization
step is done. Defaults to \code{TRUE}.}
}
\value{
List of eight items
  \item{beta.est}{Coefficient estimates}
  \item{MSPEKrig}{Mean squared Kriging Prediction error}
  \item{spatCov}{Theta parameters for stochastic term}
  \item{sigma2s}{Static sensor variability}
  \item{sigma2r}{Roving sensor variability}
  \item{static.loadings}{Loadings for the static sensors}
  \item{phi}{Estimates of the eigenfunctions evaluated at unique
             observed times in the dataset, sorted}
  \item{times}{sorted unique observed times in the dataset}
}
\description{
This function is an wrapper to fitting a Spatio-Temporal Data
Fusion (STDF) model to spatio-temporal data.
}
\examples{
# sensor simulation example, placeholder for unit tests but also shows
# options available for optimization:
set.seed(1)
static <- cbind(rep(0, 200), rep(1:50,4),
                rep(c(2,2,4,4), each = 50),
                rep(c(2,4,2,4), each = 50))
roving <- cbind(rep(0, 50), 1:50,
                seq(2,4, length=50),
                seq(2,4, length=50))
prediction <- cbind(rep(0, 50), 1:50, 3, 3)
complete <- rbind(static, roving, prediction)
D <- matrix(0, 300, 300)
for(i in 1:300) D[i,] <- sqrt((complete[i,3] - complete[,3])^2 +
                              (complete[i,4] - complete[,4])^2)
S <- matrix(0, 300, 300)
Phi <- cbind(sin(2*pi*complete[,2]/50),
             cos(2*pi*complete[,2]/50),
             2*sin(4*pi*complete[,2]/50))
sigmaR <- 2
sigmaS <- 1
theta <- 1:3
for(L in 1:3) S <- S + exp(-D/theta[L])*outer(Phi[,L], Phi[,L])
diag(S) <- diag(S) + sigmaS
diag(S)[201:250] <- diag(S)[201:250] - sigmaS + sigmaR
complete[,1] <- 2 + 2*complete[,3] + 2*complete[,4] +
                2*cos(4*pi*complete[,2]/50) +
                t(chol(S)) \%*\% rnorm(300)
static[,1] <- complete[1:200,1]
roving[,1] <- complete[201:250,1]
prediction[,1] <- complete[251:300,1]

results <- stdf(static, prediction, ssensors = 4, L = 3)
resultsNM <- stdf(static, prediction, ssensors = 4, L = 3,
                  method = "Nelder-Mead")
# Compare results of using "L-BFGS-B" and "Nelder-Mead";
# Should be almost the same, bar numberical accuracy issues
results
resultsNM
# Compare with fitting homogeneous case
resultsH <- stdf(static, prediction, ssensors = 4, L = 3,
                 homogeneous = TRUE)
resultsHNM <- stdf(static, prediction, ssensors = 4, L = 3,
                   homogeneous = TRUE, method = "Nelder-Mead")

# Nelder-Mead is more time-consuming, for example in this case:
system.time({resultsPlusRoving <- stdf(rbind(static, roving), prediction,
                          subtfpca = c(rep(TRUE,200), rep(FALSE,50)),
                          ssensors = 4, L = 3)})
system.time({resultsPlusRovingNM <- stdf(rbind(static, roving), prediction,
                            subtfpca = c(rep(TRUE,200), rep(FALSE,50)),
                            ssensors = 4, L = 3, method = "Nelder-Mead")})
# Results should be almost the same, bar numerical accuracy issues
resultsPlusRoving
resultsPlusRovingNM
# Forces homogeneous model:
resultsPlusRovingH <- stdf(rbind(static, roving), prediction,
                          subtfpca = c(rep(TRUE,200), rep(FALSE,50)),
                          ssensors = 4, L = 3, homogeneous = TRUE)

# Example with dataset from fda package; this dataset is
# huge and takes a long time to run.
\dontrun{
library(fda)
n <- 35 # stations
Y <- as.numeric(CanadianWeather$dailyAv[ , ,"Temperature.C"])
XY <- cbind(Y,
            rep(1:365, n),
            rep(CanadianWeather$coordinates[, "N.latitude"], each = 365),
            rep(CanadianWeather$coordinates[, "W.longitude"], each = 365))
fakePred <- XY[1:3,]
model <- stdf(XY, fakePred, ssensors = n, homogeneous = TRUE,
              L = 2, fpca.lambda = 1e5)
# Creating a hazard map requires the ggplot2 package
# TODO
}
}
\author{
Guilherme Ludwig and Tingjin Chu
}
\references{
Chu, T., Zhu, J. and Wang, H. (2014) On Semiparametric Inference of Geostatistical Models via Local Karhunen-Loeve Expansion. \emph{Journal of the Royal Statistical Society}, 76, 817-832.

  Ludwig, G., Chu, T., Zhu, J., Wang, H. and Koehler, K. (2015) A Geostatistical Functional Data Analysis Approach to Combining Static and Roving Sensor Data for Environmental Hazard Mapping, \emph{to appear}.
}
\seealso{
\code{\link{tfpca}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

