% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tfpca}
\alias{tfpca}
\title{Temporal Functional Principal Component}
\usage{
tfpca(MatY, J, t.fit, t.pred, tbas)
}
\arguments{
\item{MatY}{The data organized in a matrix, with each time series in a column. Missing
values are currently not handled properly.}

\item{J}{The number of functional principal components to be returned.}

\item{t.fit}{Values of the time-domain at which a value is fitted.}

\item{t.pred}{Values of the time-domain at which a predicted value needs to be produced.}

\item{tbas}{Number of spline basis functions used in the smooth.}
}
\value{
List of three items
  \item{values}{numeric vector of J Eigenvalue estimates}
  \item{vectors}{matrix of J columns with Eigenfunction estimates observed at t.fit}
  \item{pred.vec}{matrix of J columns with Eigenfunction estimates observed at t.pred}
}
\description{
This function is used on step II of the GFDA algorithm by Chu, Zhu and Wang (2014). It does
not need to be used directly, since it's mostly an wrapper to function calls from J. O. Ramsay
et al.'s \code{fda} package function calls.
}
\examples{
mean(rnorm(20)) # Write some test cases!
}
\references{
Ramsay, J. O. and Silverman, B. (2005) Functional Data Analysis. New York: Springer.
}
\seealso{
\code{\link{median}},
  \code{\link{mean}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

