% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfda}
\alias{gfda}
\title{GFDA model}
\usage{
gfda(NoiTR, NoiTE, testset, trainset, ssensors = 6, verbose = TRUE)
}
\arguments{
\item{NoiTR}{A spatio-temporal data model with vertical alignment (TODO: Generalize)}

\item{NoiTE}{A spatio-temporal data model with vertical alignment, prediction set (TODO: Generalize)}

\item{testset}{TODO}

\item{trainset}{TODO}

\item{ssensors}{How many sensors in the dataset (To be removed)}

\item{J}{Number of eigenfunctions in spatio-temporal covariance, defaults to 2 (TODO: implement)}
}
\value{
List of four items
  \item{beta.est}{Coefficient estimates}
  \item{fitted.values}{TODO}
  \item{MSPEKrig}{Mean squared Kriging Prediction error}
  \item{Cov.matrix}{TODO}
}
\description{
This function is an wrapper to fitting a Geostatistical Functional Data
Analysis (GFDA) model to spatio-temporal data.
}
\examples{
mean(rnorm(20))
}
\references{
\url{http://www.google.com}
}
\seealso{
\code{\link{median}},
  \code{\link{mean}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

