% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gfda.R
\name{gfda}
\alias{gfda}
\title{GFDA model}
\usage{
gfda(training.set, prediction.set, subtfpca = NULL, ssensors = 6, L = 2,
  spline.df = NULL, fpca.df = 10, homogeneous = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{training.set}{A numeric matrix with the response variable in the first column,
time values in the second column and X, Y coordinates in the third
column. Corresponds to the training data.}

\item{prediction.set}{Same as training.set, but corresponding to points to be predicted for
MSPE calculation.}

\item{subtfpca}{Logical vector with the same length as the number of rows in training.set; tells
gfda which lines should be included in the calculation of the temporal dependency
component. Defaults to NULL, which includes all lines.}

\item{ssensors}{Integer corresponding to how many static sensors in the dataset.}

\item{L}{Number of eigenfunctions in spatio-temporal covariance, defaults to 2.}

\item{spline.df}{Number of spline basis for the deterministic spline component (see
\code{\link{bs}} function).}

\item{fpca.df}{Number of spline basis for the stochastic spline component (see \code{\link{tfpca}}
function).}

\item{homogeneous}{Whether the variance of static and roving sensors is assumed to be the same or not.
Defaults to not.}

\item{verbose}{Whether \code{gfda} prints information about the optimization step, defaults to
               \code{TRUE}.}
}
\value{
List of four items
  \item{beta.est}{Coefficient estimates}
  \item{MSPEKrig}{Mean squared Kriging Prediction error}
  \item{spatCov}{Theta parameters for stochastic term}
  \item{sigma2s}{Static sensor variability}
  \item{sigma2r}{Roving sensor variability}
}
\description{
This function is an wrapper to fitting a Geostatistical Functional Data
Analysis (GFDA) model to spatio-temporal data.
}
\examples{
## Y <- CanadianWeather$dailyAv
## XY <- CanadianWeather$coordinates
# =sensor simulation example:
library(fda)
set.seed(1)
static <- cbind(rep(0, 200), rep(1:50,4), rep(c(2,2,4,4), each = 50), rep(c(2,4,2,4), each = 50))
roving <- cbind(rep(0, 50), 1:50, seq(2,4, length=50), seq(2,4, length=50))
prediction <- cbind(rep(0, 50), 1:50, 3, 3)
complete <- rbind(static, roving, prediction)
D <- matrix(0, 300, 300)
for(i in 1:300) D[i,] <- sqrt((complete[i,3] - complete[,3])^2 + (complete[i,4] - complete[,4])^2)
S <- matrix(0, 300, 300)
Phi <- cbind(sin(2*pi*complete[,2]/50), cos(2*pi*complete[,2]/50), 2*sin(4*pi*complete[,2]/50))
sigmaR <- 2
sigmaS <- 1
theta <- 1:3
for(L in 1:3) S <- S + exp(-D/theta[L])*outer(Phi[,L], Phi[,L])
diag(S) <- diag(S) + sigmaS
diag(S)[201:250] <- diag(S)[201:250] - sigmaS + sigmaR
complete[,1] <- 2 + 2*complete[,3] + 2*complete[,4] + 2*cos(4*pi*complete[,2]/50) +
                t(chol(S)) \%*\% rnorm(300)
static[,1] <- complete[1:200,1]
roving[,1] <- complete[201:250,1]
prediction[,1] <- complete[251:300,1]

results <- gfda(static, prediction, ssensors = 4, L = 3)
resultsH <- gfda(static, prediction, ssensors = 4, L = 3, homogeneous = TRUE)
resultsPlusRoving <- gfda(rbind(static, roving), prediction,
                          subtfpca = c(rep(TRUE,200), rep(FALSE,50)), ssensors = 4, L = 3)
}
\references{
\url{http://www.google.com}
}
\seealso{
\code{\link{tfpca}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

