% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfda}
\alias{gfda}
\title{GFDA model}
\usage{
gfda(NoiTR, NoiTE, testset, trainset, ssensors = 6, verbose = TRUE)
}
\arguments{
\item{sensor}{A spatio-temporal data model with vertical}

\item{formula}{A formula estabilishing the linear relationship component for spatial mean}

\item{SENSORS}{How many sensors in the dataset (To be removed)}

\item{J}{Number of eigenfunctions in spatio-temporal covariance, defaults to 2}

\item{method}{The fitting method, defaults to "FT" (fixed time), also allows "GFD" and "SPTLM", which implement the GFDA model (so far they're equivalent)}

\item{grid}{A choice of grid to evaluate the Noise Map. Defaults to NULL, which produces no noise map evaluation}
}
\value{
List of two items
  \item{fit}{Fitted values at the original locations}
  \item{noiseMap}{Estimated noise map}
}
\description{
This function is an wrapper to fitting a Geostatistical Functional Data
Analysis (GFDA) model to spatio-temporal data.
}
\examples{
mean(rnorm(20))
}
\references{
\url{http://www.google.com}
}
\seealso{
\code{\link{median}},
  \code{\link{mean}}
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}

